+++ src/main/java/shootingstar/var/dto/res/GetBannerResDto.java
@@ -1,0 +1,11 @@
+package shootingstar.var.dto.res;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class GetBannerResDto {
+    private String bannerImgUrl;
+    private String targetUrl;
+}
